+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the possibility to have ray traced colored and semi-transparent shadows on directional lights.
 - Added a check in the custom post process template to throw an error if the default shader is not found.
 - Exposed the debug overlay ratio in the debug menu.
+- Added a separate frame settings for tonemapping alongside color grading.
 
 ### Fixed
 - Sorting, undo, labels, layout in the Lighting Explorer.